+++ app/src/main/java/dev/saurabhmishra/exoplayersample/network/model/VideoResponse.kt
@@ -8,7 +8,7 @@ data class VideoResponse(
     val videoUrl: String? = null,
     val title: String? = null,
     val views: Long? = null,
-    val likeCount: Int? = null,
-    val dislikeCount: Int? = null,
+    val likeCount: Long? = null,
+    val dislikeCount: Long? = null,
     val uploadedTimeStamp: Long? = null
 )
